pyproject.toml
[project]
name = "tsbot"
version = "0.1.0"
description = "Chatbot AI Tư Vấn Tuyển Sinh Quân Sự Việt Nam - Agentic RAG Multi-Agent System"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "TSBot Team" }]

dependencies = [
    # Core Framework
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "websockets>=13.0",

    # Database
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "qdrant-client>=1.12.0",
    "pgvector>=0.3.0",

    # LLM & Embeddings
    "ollama>=0.4.0",
    "sentence-transformers>=3.3.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",

    # Utils
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",

    # Vietnamese NLP
    "underthesea>=6.8.0",
    "pyvi>=0.1.1",

    # Data Processing
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "pypdf>=5.0.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Logging & Monitoring
    "structlog>=24.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "httpx>=0.27.0",
]

eval = [
    "ragas>=0.2.0",
    "datasets>=3.0.0",
]

[project.scripts]
tsbot = "src.api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
